% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{get_history}
\alias{get_history}
\title{Get history}
\usage{
get_history(page = 1, pageSize = 10)
}
\arguments{
\item{page}{1-index page.}

\item{pageSize}{Number of items per page.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get history
}
\examples{
\dontrun{
history <- get_history()
}
}
