% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{get_movies}
\alias{get_movies}
\title{Get All Radarr Movies}
\source{
https://radarr.video/docs/api/#/Movie/getMovie
}
\usage{
get_movies(tmdbId = NULL)
}
\arguments{
\item{tmdbId}{\verb{[NULL]} tmdb ID of the movie to get. Default gets all movies.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get All Radarr Movies
}
\examples{
\dontrun{
movies <- get_movies()
}
}
