% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{radarr_get}
\alias{radarr_get}
\title{Make a GET request to the Radarr API}
\usage{
radarr_get(path = "/api/system/status", ...)
}
\arguments{
\item{path}{The API path, e.g. \code{"/api/system/status"} (default).}

\item{...}{Named arguments used as query parameters.}
}
\value{
A list.
}
\description{
Make a GET request to the Radarr API
}
\examples{
\dontrun{
radarr_get()
}

}
