% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarrr-package.R
\docType{package}
\name{radarrr-package}
\alias{radarrr}
\alias{radarrr-package}
\title{radarrr: Work With The 'Radarr' API}
\description{
A wrapper for the 'Radarr' API to access your
    movies and metadata.
}
\author{
\strong{Maintainer}: Lukas Burk \email{github@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
