% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_status}
\alias{system_status}
\title{Get Radarr's System Status}
\source{
https://github.com/Radarr/Radarr/wiki/API:System-Status
}
\usage{
system_status()
}
\value{
A \code{list}
}
\description{
Get Radarr's System Status
}
\examples{
\dontrun{
system_status()
}
}
