% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{get_movies}
\alias{get_movies}
\title{Get All Radarr Movies}
\source{
https://github.com/Radarr/Radarr/wiki/API:Movie
}
\usage{
get_movies(unpack = TRUE, progress = TRUE)
}
\arguments{
\item{unpack}{If \code{TRUE}(default), result is unpacked to return a flat \code{tibble}.}

\item{progress}{If \code{TRUE} (default) and \code{unpack == TRUE},
displays a progress bar via \link[dplyr:progress_estimated]{dplyr::progress_estimated} during unpacking.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of all movies
}
\description{
Get All Radarr Movies
}
\examples{
\dontrun{
movies <- get_movies()
}
}
