% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{get_movies}
\alias{get_movies}
\title{Get All Radarr Movies}
\usage{
get_movies(url = Sys.getenv("radarr_url"),
  apikey = Sys.getenv("radarr_apikey"))
}
\arguments{
\item{url}{The Radarr base URL, defaults to \code{Sys.getenv("radarr_url")}.}

\item{apikey}{The Radarr API key, defaults to \code{Sys.getenv("radarr_apikey")}}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of all movies
}
\description{
Get All Radarr Movies
}
\examples{
\dontrun{
movies <- get_movies()
}
}
