% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{system_status}
\alias{system_status}
\title{Get Radarr's System Status}
\usage{
system_status(
  url = Sys.getenv("radarr_url"),
  apikey = Sys.getenv("radarr_apikey")
)
}
\arguments{
\item{url}{The Radarr base URL, defaults to \code{Sys.getenv("radarr_url")}.}

\item{apikey}{The Radarr API key, defaults to \code{Sys.getenv("radarr_apikey")}}
}
\value{
A \code{list}
}
\description{
Get Radarr's System Status
}
\examples{
\dontrun{
system_status()
}
}
